Add `AccountRootPrincipal` ARN as KMS key user
This enables IAM policies to allow access to the CMK used for snapshot
export encryption, such as for the account administrator or a user
assuming an authorized IAM role. As is the case with the default KMS
key policy, this alone does not grant access to the key itself.

See: https://docs.aws.amazon.com/kms/latest/developerguide/key-policies.html#key-policy-default